Split pagination components into separate files
