+++ packages/palette/src/elements/Pagination/LargePagination.tsx
@@ -1,13 +1,12 @@
 import React from "react"
 import styled, { css } from "styled-components"
 import { color } from "../../helpers/color"
-import { space } from "../../helpers/space"
 import { ChevronIcon } from "../../svgs/ChevronIcon"
 import { Box } from "../Box"
 import { Flex } from "../Flex"
 import { Sans } from "../Typography"
 
-interface Props {
+export interface Props {
   onClick?: (cursor: string, page: number) => void
   onNext?: () => void
   pageCursors: any
@@ -15,7 +14,6 @@ interface Props {
   scrollTo?: string
 }
 
-/** LargePagination */
 export const LargePagination = (props: Props) => {
   const {
     pageCursors: { around, first, last, previous },
@@ -62,61 +60,6 @@ export const LargePagination = (props: Props) => {
   )
 }
 
-/** SmallPagination */
-export const SmallPagination = (props: Props) => {
-  const {
-    pageCursors: { previous },
-    onClick,
-    onNext,
-    hasNextPage,
-  } = props
-
-  return (
-    <Flex flexDirection="row" width="100%">
-      <PrevNextFlex
-        className={!previous ? "disabled" : null}
-        width="50%"
-        pr={0.5}
-      >
-        <ButtonWithBorder
-          alignItems="center"
-          justifyContent="flex-start"
-          pl={1}
-          onClick={() => {
-            if (previous) {
-              onClick(previous.cursor, previous.page)
-            }
-          }}
-        >
-          <ChevronIcon direction="left" />
-        </ButtonWithBorder>
-      </PrevNextFlex>
-      <PrevNextFlex
-        className={!hasNextPage ? "disabled" : null}
-        width="50%"
-        pl={0.5}
-      >
-        <ButtonWithBorder
-          onClick={() => onNext()}
-          alignItems="center"
-          justifyContent="flex-end"
-          pr={1}
-        >
-          <ChevronIcon direction="right" />
-        </ButtonWithBorder>
-      </PrevNextFlex>
-    </Flex>
-  )
-}
-
-const ButtonWithBorder = styled(Flex)`
-  border: ${props => props.theme.borders[1]};
-  border-color: ${color("black10")};
-  border-radius: 3px;
-  width: 100%;
-  height: ${space(4)}px;
-  cursor: pointer;
-`
 const PageSpan = ({ mx }) => {
   return (
     <Sans size="3" display="inline" mx={mx} color="black30">
@@ -171,6 +114,7 @@ const Button = styled.button<{ active?: boolean }>`
     ${activeButton};
   }
 `
+
 const PrevButton = ({ onClick, disabled }) => {
   return (
     <PrevNextContainer className={disabled ? "disabled" : null}>
@@ -201,13 +145,6 @@ const PrevNextContainer = styled.span`
   }
 `
 
-const PrevNextFlex = styled(Flex)`
-  &.disabled {
-    opacity: 0.1;
-  }
-`
-
 // Tests
-ButtonWithBorder.displayName = "ButtonWithBorder"
 PrevButton.displayName = "PrevButton"
 NextButton.displayName = "NextButton"
